refactor(344): reorganize NewsletterSubscriber type imports and exports 2
