+++ src/types/newsletterSubscriber.types.ts
